update and delete by the user connected

+++ backend/controllers/websiteController.js
@@ -6,19 +6,24 @@ const Website = require("../models/websiteModel");
 // Route GET /api/web
 // Aces Public
 const getWebsite = asyncHandler(async (req, res) => {
-  const websites = await Website.find();
+  const website = await Website.find({ user: req.user.id });
 
-  res.status(200).json(websites);
+  res.status(200).json(website);
 });
 
 // Post website
 // Route Post /api/web
 // Aces Private
 const postWebsite = asyncHandler(async (req, res) => {
-  const websites = await Website.create({
+  if (!req.body.text) {
+    res.status(400);
+    throw new Error("please add a text field");
+  }
+  const website = await Website.create({
     text: req.body.text,
+    user: req.user.id,
   });
-  res.status(200).json(websites);
+  res.status(200).json(website);
 });
 
 // Update website
@@ -29,6 +34,18 @@ const updateWebsite = asyncHandler(async (req, res) => {
   if (!website) {
     res.status(400);
     throw new Error("website was not found");
+  }
+
+  //check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Make sure the logged in user matches the profile user
+  if (website.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   const updatedWebsite = await Website.findByIdAndUpdate(
@@ -49,6 +66,18 @@ const deleteWebsite = asyncHandler(async (req, res) => {
   if (!website) {
     res.status(400);
     throw new Error("website was not found");
+  }
+
+  //check for user
+  if (!req.user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Make sure the logged in user matches the profile user
+  if (website.user.toString() !== req.user.id) {
+    res.status(401);
+    throw new Error("User not authorized");
   }
 
   await website.deleteOne();
